Finished the remaining keybinds
