+++ src/Sidekick.Application/Game/Chat/LeavePartyHandler.cs
@@ -1,36 +1,48 @@
+using System.Threading;
 using System.Threading.Tasks;
 using MediatR;
 using Microsoft.Extensions.Logging;
 using Sidekick.Domain.Game.Chat.Commands;
+using Sidekick.Domain.Process;
 using Sidekick.Domain.Settings;
 
-namespace Sidekick.Business.Parties
+namespace Sidekick.Application.Game.Chat
 {
-    public class PartyService : IPartyService
+    public class LeavePartyHandler : ICommandHandler<LeavePartyCommand, bool>
     {
-        private readonly ILogger logger;
         private readonly ISidekickSettings settings;
+        private readonly ILogger<LeavePartyHandler> logger;
         private readonly IMediator mediator;
+        private readonly INativeProcess nativeProcess;
 
-        public PartyService(
-            ILogger<PartyService> logger,
+        public LeavePartyHandler(
             ISidekickSettings settings,
-            IMediator mediator)
+            ILogger<LeavePartyHandler> logger,
+            IMediator mediator,
+            INativeProcess nativeProcess)
         {
-            this.logger = logger;
             this.settings = settings;
+            this.logger = logger;
             this.mediator = mediator;
+            this.nativeProcess = nativeProcess;
         }
 
-        public async Task LeaveParty()
+        public async Task<bool> Handle(LeavePartyCommand request, CancellationToken cancellationToken)
         {
+            if (!nativeProcess.IsPathOfExileInFocus)
+            {
+                return false;
+            }
+
             // This operation is only valid if the user has added their character name to the settings file.
             if (string.IsNullOrEmpty(settings.Character_Name))
             {
                 logger.LogWarning(@"This command requires a ""CharacterName"" to be specified in the settings menu.");
-                return;
+                return false;
             }
+
             await mediator.Send(new WriteChatCommand($"/kick {settings.Character_Name}"));
+            return true;
         }
     }
 }
