+++ src/Sidekick.Application/Game/Chat/ReplyToLatestWhisperHandler.cs
