--- src/Sidekick.Business/Parties/IPartyService.cs
@@ -1,9 +1,0 @@
-﻿using System.Threading.Tasks;
-
-namespace Sidekick.Business.Parties
-{
-    public interface IPartyService
-    {
-        Task LeaveParty();
-    }
-}
