+++ src/Sidekick.Domain/Wikis/Commands/OpenWikiPageCommand.cs
@@ -1,0 +1,11 @@
+using MediatR;
+
+namespace Sidekick.Domain.Wikis.Commands
+{
+    /// <summary>
+    /// Opens a webpage to the wiki of the item under the cursor inside Path of Exile
+    /// </summary>
+    public class OpenWikiPageCommand : ICommand<bool>
+    {
+    }
+}
